IS-3304: Bare nulle oppfolgingsenhet when veileder (#553)

+++ src/main/kotlin/no/nav/syfo/personstatus/infrastructure/cronjob/reaper/ReaperCronjob.kt
@@ -39,12 +39,14 @@ class ReaperCronjob(
             .forEach {
                 try {
                     val personident = PersonIdent(it.fnr)
-                    personOversiktStatusService.removeTildeltVeileder(personident)
                     personOversiktStatusService.removeTildeltEnhet(personident)
-                    behandlendeEnhetClient.unsetOppfolgingsenhet(
-                        callId = UUID.randomUUID().toString(),
-                        personIdent = personident,
-                    )
+                    if (it.veilederIdent != null) {
+                        personOversiktStatusService.removeTildeltVeileder(personident)
+                        behandlendeEnhetClient.unsetOppfolgingsenhet(
+                            callId = UUID.randomUUID().toString(),
+                            personIdent = personident,
+                        )
+                    }
                     result.updated++
                     COUNT_CRONJOB_REAPER_UPDATE.increment()
                 } catch (ex: Exception) {
